refactor: - Campos no banco de dados referentes a valores monetários sofreram uma alteração no tipo, de BIGINT fora para NUMERIC (valores reais)

+++ GestaoDeBarbearia.Communication/Responses/ResponseScheduledServiceJson.cs
@@ -2,6 +2,6 @@
 public class ResponseScheduledServiceJson
 {
     public DateTime AppointmentDateTime { get; set; }
-    public long ServicePrice { get; set; }
+    public decimal ServicePrice { get; set; }
     public long? EmployeeId { get; set; }
 }
