+++ GestaoDeBarbearia.Domain/Entities/Appointment.cs
@@ -10,7 +10,7 @@ public class Appointment
     public string? ClientPhone { get; set; }
     public long? EmployeeId { get; set; }
     public AppointmentStatus Status { get; set; }
-    public long ServicePrice { get; set; }
+    public decimal ServicePrice { get; set; }
     public PaymentType PaymentType { get; set; }
     public DateTime? PaidAt { get; set; }
     public DateTime CreatedAt { get; set; }
