+++ GestaoDeBarbearia.Domain/Entities/Employee.cs
@@ -8,7 +8,7 @@ public class Employee
     public string Telephone { get; set; } = string.Empty;
     public string Email { get; set; } = string.Empty;
     public string Password { get; set; } = string.Empty;
-    public long Salary { get; set; }
+    public decimal Salary { get; set; }
     public EmployeePosition Position { get; set; }
     public DateTime CreatedAt { get; set; }
     public DateTime UpdatedAt { get; set; }
