+++ GestaoDeBarbearia.Domain/Entities/Product.cs
@@ -3,8 +3,8 @@ public class Product
 {
     public long Id { get; set; }
     public string Name { get; set; } = string.Empty;
-    public long SalePrice { get; set; }
-    public long UnitCost { get; set; }
+    public decimal SalePrice { get; set; }
+    public decimal UnitCost { get; set; }
     public long Quantity { get; set; }
     public int MinimumStock { get; set; }
     public DateTime CreatedAt { get; set; }
